refactor: test layer과 product layer 분리 및 FixedIssueStrategy 테스트코드 삭제
